#1421 moved metahist read to smvApp
